Add grade list
add front end files for grade list
frontend is fetching grades from backend with hardcoded assignmentId.
Add a nav buttom for loading the grades fragment directly (for testing and demo purposes, remove later).

+++ front_end/app/src/main/java/com/example/studentportal/grades/ui/model/GradeUiModel.kt
@@ -4,7 +4,7 @@ import com.example.studentportal.common.ui.model.BaseUiModel
 
 data class GradeUiModel(
     val gradeId: String,
-    val score: Double,
+    val score: Int,
     val studentFirstName: String,
     val studentLastName: String,
     val studentId: String,
